Initial implementation for min and max intervals

+++ src/main/java/com/texoit/exercises/awardsapi/domain/entity/Movie.java
@@ -3,6 +3,8 @@ package com.texoit.exercises.awardsapi.domain.entity;
 import jakarta.persistence.*;
 import lombok.Data;
 
+import java.util.List;
+
 @Data
 @Entity
 public class Movie {
@@ -10,7 +12,7 @@ public class Movie {
     @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "movie_generator")
     @SequenceGenerator(name = "movie_generator", sequenceName = "MOVIE_SEQ", allocationSize = 1)
     private Long movieId;
-    private int releaseYear;
+    private Long releaseYear;
     private String title;
     private String studios;
     private String producers;
